requirements.txt
requests
python-dotenv # for environment var management
pylint